% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestclean-description.R
\docType{package}
\name{nestclean-package}
\alias{nestclean}
\alias{nestclean-package}
\title{nestclean: A Package for Checking NEST Lab Data Cleaning Projects}
\description{
The `nestclean` package provides tools to check and summarize the outcome of data cleaning for the NEST lab.
}
\section{Main Functions}{

- `check_continuous()`: Summarizes continuous variables by calculating mean, minimum, maximum, and missing count.
- `check_categorical()`: Summarizes categorical variables by printing the value level, value label, and frequency.
- `inspect_labels()`: Retrieves the labels of specified variables.
- `print_slice()`: Prints a random sample of 6 rows from specified columns in a data frame (HATCH only).
- `print_slice_chirp()`: Prints a random sample of 6 rows from specified columns in a data frame (CHIRP or any df with "ID").
- `view_selected()`: Displays a specified subset of columns in a data frame (HATCH only).
- `view_selected_chirp()`: Displays a specified subset of columns in a data frame (CHIRP or any df with "ID").
}

\author{
\strong{Maintainer}: Elizabeth Aviv \email{eaviv@usc.edu}

}
\keyword{internal}
