% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_slice}
\alias{print_slice}
\title{Print a Random Slice of Data: HATCH}
\usage{
print_slice(df, ...)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{...}{One or more unquoted expressions separated by commas, indicating variables to print (e.g., column names, column ranges, or selection helpers like contains()).}
}
\value{
A formatted table displaying a random slice of the selected data.
}
\description{
This function selects and prints a random sample of 6 rows from specified columns in a data frame, formatted as a table.
}
