% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{check_continuous}
\alias{check_continuous}
\title{Summarize Continuous Variables}
\usage{
check_continuous(df, summary_cols)
}
\arguments{
\item{df}{A data frame containing the variables to be summarized.}

\item{summary_cols}{A tidy-select expression that specifies the columns to summarize.}
}
\value{
A data frame in long format with the summary statistics for each variable.
}
\description{
This function calculates summary statistics (mean, minimum, maximum, and missing count) for specified continuous variables in a data frame. The results are returned in a long format with each statistic as a separate row.
}
