% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inspect_labels}
\alias{inspect_labels}
\title{Retrieve Variable Labels}
\usage{
inspect_labels(df, columns_range)
}
\arguments{
\item{df}{A data frame containing the variables.}

\item{columns_range}{A tidy-select expression that specifies the columns to retrieve labels for.}
}
\value{
A formatted table displaying the variable labels.
}
\description{
This function retrieves the labels of specified variables in a data frame and displays them in a formatted table.
}
