% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_slice}
\alias{print_slice}
\title{Print a Random Slice of Data}
\usage{
print_slice(df, columns_range)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{columns_range}{A tidy-select expression that specifies the columns to be printed along with `CoupleID` and `Parent`.}
}
\value{
A formatted table displaying a random slice of the selected data.
}
\description{
This function selects and prints a random sample of 6 rows from specified columns in a data frame, formatted as a table.
}
