% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{view_selected}
\alias{view_selected}
\title{View Specific Columns in Data Frame}
\usage{
view_selected(df, ...)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{...}{One or more unquoted expressions separated by commas, indicating variables to view (e.g., column names, column ranges, or selection helpers like contains()).}
}
\value{
Opens the specified columns in a new View window.
}
\description{
This function displays a specified subset of columns in a data frame, including `CoupleID` and `Parent`.
}
