% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{view_selected}
\alias{view_selected}
\title{View Specific Columns in Data Frame}
\usage{
view_selected(df, columns_range)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{columns_range}{A tidy-select expression that specifies the columns to view along with `CoupleID` and `Parent`.}
}
\value{
Opens the specified columns in a new View window.
}
\description{
This function displays a specified subset of columns in a data frame, including `CoupleID` and `Parent`.
}
